+++ commercial/app/model/capi/Lookup.scala
@@ -5,6 +5,7 @@ import common.Edition.defaultEdition
 import common.Logging
 import contentapi.ContentApiClient
 import model.{Content, ContentType, ImageElement}
+import utils.ShortUrls
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -26,35 +27,11 @@ class Lookup(contentApiClient: ContentApiClient) extends Logging with implicits.
     }
   }
 
-  def shortUrlToShortId(shortUrl: String): String = {
-    /*
-        Date: 02nd Dec 2020
-        id: 288767d7-ba82-4d67-8fb3-9139e67b0f2e
-
-        CAPI (recently) announced that we would be moving
-
-        from
-          https://gu.com/p/abc
-
-        to
-          https://theguardian.com/p/abc
-
-        for "short" urls.
-
-        Introducing this to handle gracefully both the old and new convention without having to
-        worry when the change will actually happen. This function can be simplified in the future when the migration
-        has completed.
-     */
-    shortUrl
-      .replaceFirst("^[a-zA-Z]+://gu.com/", "")
-      .replaceFirst("^[a-zA-Z]+://theguardian.com/", "")
-  }
-
   def contentByShortUrls(
       shortUrls: Seq[String],
   )(implicit executionContext: ExecutionContext): Future[Seq[ContentType]] = {
     if (shortUrls.nonEmpty) {
-      val shortIds = shortUrls.map(shortUrlToShortId).mkString(",")
+      val shortIds = shortUrls.map(ShortUrls.shortUrlToShortId).mkString(",")
       contentApiClient.getResponse(contentApiClient.search(defaultEdition).ids(shortIds)) map {
         _.results map (Content(_))
       }
