+++ common/app/model/meta.scala
@@ -23,6 +23,7 @@ import play.twirl.api.Html
 import navigation.GuardianFoundationHelper
 
 import scala.util.matching.Regex
+import utils.ShortUrls
 
 object Commercial {
 
@@ -111,36 +112,7 @@ final case class Fields(
     showAffiliateLinks: Option[Boolean],
 ) {
 
-  def ensureStartingForwardSlash(str: String): String = {
-    if (!str.startsWith("/")) ("/" + str) else str
-  }
-
-  def shortUrlToShortId(shortUrl: String): String = {
-    /*
-        Date: 02nd Dec 2020
-        id: 288767d7-ba82-4d67-8fb3-9139e67b0f2e
-
-        CAPI (recently) announced that we would be moving
-
-        from
-          https://gu.com/p/abc
-
-        to
-          https://theguardian.com/p/abc
-
-        for "short" urls.
-
-        Introducing this to handle gracefully both the old and new convention without having to
-        worry when the change will actually happen. This function can be simplified in the future when the migration
-        has completed.
-     */
-    val id = shortUrl
-      .replaceFirst("^[a-zA-Z]+://gu.com/", "")
-      .replaceFirst("^[a-zA-Z]+://theguardian.com/", "")
-    ensureStartingForwardSlash(id)
-  }
-
-  lazy val shortUrlId = shortUrlToShortId(shortUrl)
+  lazy val shortUrlId = ShortUrls.shortUrlToShortIdWithStartingForwardSlash(shortUrl)
   lazy val isRightToLeftLang: Boolean = lang.contains("ar")
 
   def javascriptConfig: Map[String, JsValue] = {
